+++ device-bootloaders/src/ServiceProcessorBL2/include/bl2_pmic_controller.h
@@ -207,12 +207,12 @@ int pmic_set_temperature_threshold(uint8_t temp_limit);
 */
 int pmic_get_temperature(uint8_t *sys_temp);
 
-/*! \fn int pmic_read_instantaneous_soc_power(uint8_t* soc_pwr)
+/*! \fn int pmic_read_instantaneous_soc_power(uint16_t* soc_pwr)
     \brief This function returns soc input power.
     \param soc_pwr - value of Input Power (binary encoded).
     \return The function call status, pass/fail.
 */
-int pmic_read_instantaneous_soc_power(uint8_t *soc_pwr);
+int pmic_read_instantaneous_soc_power(uint16_t *soc_pwr);
 
 /*! \fn int pmic_enable_etsoc_reset_after_perst(void)
     \brief This function enables a PERST event to also reset the ET-SOC.
@@ -376,12 +376,12 @@ int pmic_force_perst(void);
 */
 int pmic_reset_wdog_timer(void);
 
-/*! \fn int pmic_read_average_soc_power(uint8_t* avg_power)
+/*! \fn int pmic_read_average_soc_power(uint16_t* avg_power)
     \brief This function gets Average Power.
     \param avg_power - value of Average Power (binary encoded).
     \return The function call status, pass/fail.
 */
-int pmic_read_average_soc_power(uint8_t *avg_power);
+int pmic_read_average_soc_power(uint16_t *avg_power);
 
 /*! \fn int I2C_PMIC_Initialize(uint8_t i2c_id)
     \brief This function initializes PMIC
@@ -410,20 +410,6 @@ int I2C_PMIC_Write(uint8_t reg, uint8_t data);
     \return The function call status, pass/fail
 */
 int pmic_get_int_config(uint8_t *int_config);
-
-/*! \fn Power_Convert_Hex_to_mW(int8_t power_hex) 
-    \brief This function converts PMIC encoded HEX value to real Power(mW)
-    \param Hex  Power in PMIC encoded Hex value 
-    \return Power in mW after conversion
-*/
-int32_t Power_Convert_Hex_to_mW(uint8_t power_hex);
-
-/*! \fn Power_Convert_mW_to_Hex(int8_t power_mW) 
-    \brief This function converts real Power(mW) to PMIC encoded HEX value
-    \param Power in mW after conversion
-    \return Hex  Power in PMIC encoded Hex value 
-*/
-int32_t Power_Convert_mW_to_Hex(uint8_t power_mW);
 
 /*! \fn int pmic_get_board_type(uint32_t *board_type)
     \brief This function reads board type.