+++ device-bootloaders/src/test_framework/tf_pmic_tests.c
@@ -3,14 +3,14 @@
 #include "bl2_pmic_controller.h"
 #include "thermal_pwr_mgmt.h"
 
-int8_t PMIC_Module_Temperature_Cmd_Handler(void* test_cmd);
-int8_t PMIC_Module_Power_Cmd_Handler(void* test_cmd);
-int8_t PMIC_Module_Voltage_Cmd_Handler(const void* test_cmd);
-int8_t PMIC_Module_Uptime_Cmd_Handler(const void* test_cmd);
+int8_t PMIC_Module_Temperature_Cmd_Handler(void *test_cmd);
+int8_t PMIC_Module_Power_Cmd_Handler(void *test_cmd);
+int8_t PMIC_Module_Voltage_Cmd_Handler(const void *test_cmd);
+int8_t PMIC_Module_Uptime_Cmd_Handler(const void *test_cmd);
 
-int8_t PMIC_Module_Temperature_Cmd_Handler(void* test_cmd)
+int8_t PMIC_Module_Temperature_Cmd_Handler(void *test_cmd)
 {
-    (void) test_cmd;
+    (void)test_cmd;
     struct tf_rsp_pmic_module_temperature_t rsp = { 0 };
 
     rsp.rsp_hdr.id = TF_RSP_PMIC_MODULE_TEMPERATURE;
@@ -24,9 +24,10 @@ int8_t PMIC_Module_Temperature_Cmd_Handler(void* test_cmd)
     return 0;
 }
 
-int8_t PMIC_Module_Power_Cmd_Handler(void* test_cmd)
+int8_t PMIC_Module_Power_Cmd_Handler(void *test_cmd)
 {
-    (void) test_cmd;
+    (void)test_cmd;
+
     struct tf_rsp_pmic_module_power_t rsp = { 0 };
 
     rsp.rsp_hdr.id = TF_RSP_PMIC_MODULE_POWER;
@@ -34,24 +35,22 @@ int8_t PMIC_Module_Power_Cmd_Handler(void* test_cmd)
     rsp.rsp_hdr.payload_size = TF_GET_PAYLOAD_SIZE(rsp);
 
     pmic_read_instantaneous_soc_power(&rsp.mod_power);
-
     TF_Send_Response(&rsp, sizeof(rsp));
 
     return 0;
 }
 
-int8_t PMIC_Module_Voltage_Cmd_Handler(const void* test_cmd)
+int8_t PMIC_Module_Voltage_Cmd_Handler(const void *test_cmd)
 {
-    (void) test_cmd;
-    struct module_voltage_t module_voltage ={0};
+    (void)test_cmd;
+    struct module_voltage_t module_voltage = { 0 };
     struct tf_rsp_pmic_module_voltage_t rsp;
 
     rsp.rsp_hdr.id = TF_RSP_PMIC_MODULE_VOLTAGE;
     rsp.rsp_hdr.flags = TF_RSP_WITH_PAYLOAD;
-    rsp.rsp_hdr.payload_size =
-                TF_GET_PAYLOAD_SIZE(struct tf_rsp_pmic_module_voltage_t);
+    rsp.rsp_hdr.payload_size = TF_GET_PAYLOAD_SIZE(struct tf_rsp_pmic_module_voltage_t);
 
-    if( 0 == get_module_voltage(&module_voltage))
+    if (0 == get_module_voltage(&module_voltage))
     {
         rsp.ddr_voltage = module_voltage.ddr;
         rsp.l2_cache_voltage = module_voltage.l2_cache;
@@ -67,18 +66,17 @@ int8_t PMIC_Module_Voltage_Cmd_Handler(const void* test_cmd)
     return TF_Send_Response(&rsp, sizeof(struct tf_rsp_pmic_module_voltage_t));
 }
 
-int8_t PMIC_Module_Uptime_Cmd_Handler(const void* test_cmd)
+int8_t PMIC_Module_Uptime_Cmd_Handler(const void *test_cmd)
 {
-    (void) test_cmd;
-    struct module_uptime_t module_uptime = {0};
+    (void)test_cmd;
+    struct module_uptime_t module_uptime = { 0 };
     struct tf_rsp_pmic_module_uptime_t rsp;
 
     rsp.rsp_hdr.id = TF_RSP_PMIC_MODULE_UPTIME;
     rsp.rsp_hdr.flags = TF_RSP_WITH_PAYLOAD;
-    rsp.rsp_hdr.payload_size =
-                TF_GET_PAYLOAD_SIZE(struct tf_rsp_pmic_module_uptime_t);
+    rsp.rsp_hdr.payload_size = TF_GET_PAYLOAD_SIZE(struct tf_rsp_pmic_module_uptime_t);
 
-    if(0 == get_module_uptime(&module_uptime))
+    if (0 == get_module_uptime(&module_uptime))
     {
         rsp.day = module_uptime.day;
         rsp.hour = module_uptime.hours;
@@ -87,4 +85,3 @@ int8_t PMIC_Module_Uptime_Cmd_Handler(const void* test_cmd)
 
     return TF_Send_Response(&rsp, sizeof(struct tf_rsp_pmic_module_uptime_t));
 }
-